refactor models to separate location
